FetchSentinel refactoring
